+++ packages/bluesky/lib/src/services/entities/muted_word.g.dart
@@ -1,0 +1,32 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// ignore_for_file: non_constant_identifier_names
+
+part of 'muted_word.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+_$MutedWordImpl _$$MutedWordImplFromJson(Map json) => $checkedCreate(
+      r'_$MutedWordImpl',
+      json,
+      ($checkedConvert) {
+        final val = _$MutedWordImpl(
+          value: $checkedConvert('value', (v) => v as String),
+          targets: $checkedConvert(
+              'targets',
+              (v) => (v as List<dynamic>)
+                  .map((e) => MutedWordTarget.fromJson(
+                      Map<String, Object?>.from(e as Map)))
+                  .toList()),
+        );
+        return val;
+      },
+    );
+
+Map<String, dynamic> _$$MutedWordImplToJson(_$MutedWordImpl instance) =>
+    <String, dynamic>{
+      'value': instance.value,
+      'targets': instance.targets.map((e) => e.toJson()).toList(),
+    };